removing items has been added

+++ src/app/features/documents/store/documents.state.ts
@@ -10,9 +10,12 @@ import { Document } from '../../../shared/models/document';
 import { computed, inject } from '@angular/core';
 import { DocumentsService } from '../services/documents.service';
 import { rxMethod } from '@ngrx/signals/rxjs-interop';
-import { switchMap } from 'rxjs';
+import { asapScheduler, map, scheduled, switchMap } from 'rxjs';
 import { tapResponse } from '@ngrx/operators';
 import { CreateDocumentReq } from '../models/create-document-req.interface';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmationModalComponent } from '../../../shared/components/confirmation-modal/confirmation-modal.component';
 
 const initialState = {
   items: <Document[]>[],
@@ -27,6 +30,8 @@ export const DocumentsStore = signalStore(
 
   withMethods((store) => {
     const service = inject(DocumentsService);
+    const dialog = inject(MatDialog);
+    const snackBar = inject(MatSnackBar);
 
     return {
       getDocuments: rxMethod<void>(
@@ -38,7 +43,10 @@ export const DocumentsStore = signalStore(
               next: (res) => {
                 patchState(store, { items: res.results });
               },
-              error: console.error,
+              error: (error) => {
+                console.error(error);
+                snackBar.open('Something went wrong', 'Close');
+              },
               finalize: () => patchState(store, { loading: false }),
             }),
           );
@@ -54,8 +62,14 @@ export const DocumentsStore = signalStore(
               next: (item) => {
                 patchState(store, { items: [...store.items(), item] });
               },
-              error: console.error,
-              finalize: () => patchState(store, { loading: false }),
+              error: (error) => {
+                console.error(error);
+                snackBar.open('Something went wrong', 'Close');
+              },
+              finalize: () => {
+                patchState(store, { loading: false });
+                snackBar.open('Document has been created', 'Close');
+              },
             }),
           );
         }),
@@ -76,9 +90,14 @@ export const DocumentsStore = signalStore(
                   items: updatedItems,
                 });
               },
-              error: console.error,
-              finalize: () =>
-                patchState(store, { updating: false, loading: false }),
+              error: (error) => {
+                console.error(error);
+                snackBar.open('Something went wrong', 'Close');
+              },
+              finalize: () => {
+                patchState(store, { updating: false, loading: false });
+                snackBar.open('Document has been updated', 'Close');
+              },
             }),
           );
         }),
@@ -86,19 +105,39 @@ export const DocumentsStore = signalStore(
 
       deleteDocument: rxMethod<string>(
         switchMap((id) => {
-          patchState(store, { loading: true });
+          const fileName = store.items().find((item) => item.id === id)?.name;
+          return dialog
+            .open(ConfirmationModalComponent, {
+              data: fileName,
+            })
+            .afterClosed()
+            .pipe(
+              map((confirmation) => (confirmation ? id : null)),
+              switchMap((id) => {
+                if (!id) {
+                  return scheduled([undefined], asapScheduler);
+                }
+                patchState(store, { loading: true });
 
-          return service.removeDocument(id).pipe(
-            tapResponse({
-              next: () => {
-                patchState(store, {
-                  items: store.items().filter((item) => item.id !== id),
-                });
-              },
-              error: console.error,
-              finalize: () => patchState(store, { loading: false }),
-            }),
-          );
+                return service.removeDocument(id).pipe(
+                  tapResponse({
+                    next: () => {
+                      patchState(store, {
+                        items: store.items().filter((item) => item.id !== id),
+                      });
+                    },
+                    error: (error) => {
+                      console.error(error);
+                      snackBar.open('Something went wrong', 'Close');
+                    },
+                    finalize: () => {
+                      patchState(store, { loading: false });
+                      snackBar.open('Document has been deleted', 'Close');
+                    },
+                  }),
+                );
+              }),
+            );
         }),
       ),
     };
